Updated dependency versions
Also added a preinstall script to package.json to verify that libfontconfig is installed in the host OS before continuing. PhantomJS requires libfontconfig, and luckily it had been present in the host OS so far, but today I ran into a development environment where it wasn't present and it caused vague errors that took a while to track down. This should make sure that is present in the OS going forward.