Added a timeout to spawns.js
If the spawned child process takes longer than 30 seconds, the promise resolves with a message to that effect. Also added a function to clean up spawned child processes after the report is done. Had to include a hacky workaround to cleanup the orphaned phantomjs process I was seeing leftover (watching the processes by using the linux 'top' command). Also added express-enforce-ssl when running in production.